+++ ChangeLog.md
@@ -52,6 +52,7 @@
     - Add view layout
     - Add main menu in toolbaar
     - Add activity menu
+    - Add splash screen
   - Add build_mscos.sh. add App bundle
   - Add run the program directly without installation for windows install package
   - [Add store password to system redential](https://github.com/KangLin/RabbitRemoteControl/issues/78)